Update version and SEP_ALT
Changed my mind, it's better to have SEP_ALT relative to the sea level. Because if for a reason it reboots in flight, it'll take the actual altitude as ground altitude and never separate. While if we use sea level as a reference, whatever the reboot altitude is, it'll always deploy daughter at the right altitude.